+++ tests/kernel/libs/src/Makefile
@@ -1,3 +1,4 @@
+ccflags-y += -I${ZEPHYR_BASE}/tests/include
 include $(ZEPHYR_BASE)/tests/Makefile.test
 
-obj-y = libraries.o
+obj-y = main.o libraries.o
