+++ tests/kernel/libs/src/libraries.c
@@ -262,16 +262,3 @@ void string_test(void)
 	strchr_test();
 	memcmp_test();
 }
-
-void test_main(void)
-{
-
-	ztest_test_suite(test_libs,
-			 ztest_unit_test(limits_test),
-			 ztest_unit_test(stdbool_test),
-			 ztest_unit_test(stddef_test),
-			 ztest_unit_test(stdint_test),
-			 ztest_unit_test(string_test));
-
-	ztest_run_test_suite(test_libs);
-}
